chore: move /shared to /core
