+++ tests/e2e/test_client.py
@@ -6,8 +6,8 @@ import pytest
 
 from src.core import response_types
 from src.core.http_client import HTTPClient
+from src.core.types import UUID, Model, TaskStatus
 from src.domain.model import Task
-from src.shared.types import UUID, Model, TaskStatus
 from tests.e2e.conftest import TEST_SERVER_DOMAIN, TEST_SERVER_PORT
