+++ tests/integration/service_layer/test_uow.py
@@ -6,9 +6,9 @@ import pytest
 from sqlalchemy import text
 from sqlalchemy.orm import Session
 
+from src.core.types import UUID, Model, TaskStatus
 from src.domain.model import Task
 from src.service_layer.sqlalchemy_uow import SessionFactory, SQLAlchemyUoW
-from src.shared.types import UUID, Model, TaskStatus
 
 
 class CustomException(Exception):
