+++ tests/unit/service_layer/fakes.py
@@ -7,8 +7,8 @@ from random import Random
 from typing import List, Optional, Set
 
 from src.adapters.repository import AbstractRepository
+from src.core.types import UUID, Model, TaskStatus
 from src.domain.model import Task
-from src.shared.types import UUID, Model, TaskStatus
 
 
 @dataclass
